Move `$app->roles()` to `AppUsers` trait
Improves consistency and overview
